requirements.txt
# Core Django packages
Django==5.1.3
djangorestframework==3.15.2
django-filter==24.3
djangorestframework-simplejwt==5.3.1

# Image processing
Pillow==11.1.0   # Updated for Python 3.13 compatibility

# Configuration management
python-decouple==3.8

# CORS handling
django-cors-headers==4.4.0

# Development tools
django-extensions==3.2.1

# Database support (PostgreSQL for Render)
dj-database-url==2.1.0
psycopg2-binary==2.9.9

# Static files serving
whitenoise==6.6.0

# WSGI server
gunicorn==21.2.0